Make DigitsTest check for a correct toString() method.
